Windows: fix crash when rendering video
Becuase stdlib.h was included, malloc was not declared anywhere. In C,
using an undeclared function makes the compiler pretend that it's
declared and has a signature of

int func();

meaning that the return value is a 32 bit int and the parameters can be
whatever.

As a result of this, after the malloc call, the compiler would insert an
instruction to sign extend the lower 32 bits of the RAX register into
the upper 32 bits, which would of course create a pointer that is
invalid in a lot of cases